requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4